Moved utility modules for the FLASH backend into a separate folder
